Update HRManagerProfile' manager
Now using a helper to ensure permission on a model to a group. Also refactored the management command to create the hr group to use this helper. The default permissions are only set on Applications and Applicant so a newly create HEManager via create_with_new_user mgmt command will only manage those models.